payment-handler/payment-request-event.https.html

<!doctype html>
<meta charset="utf-8">
<title>Tests for PaymentRequestEvent</title>
<link rel="help" href="https://w3c.github.io/payment-handler/#the-paymentrequestevent">
<link rel="manifest" href="/payment-handler/basic-card.json">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>When the payment sheet is shown, please authorize the mock payment.</p>
<script>
// Use a different scope for the payment handler, because Chrome closes all
// payment handler scoped windows after payment.
const SCOPE = 'payment-app/';
async function installPaymentHandlerAndRunTheTests() {
  const registration = await navigator.serviceWorker.register('basic-card.js', {scope: SCOPE});
  const serviceWorker = registration.installing || registration.waiting || registration.active;
  if (serviceWorker) {
    waitForServiceWorkerActivation();
    return;
  }

  registration.addEventListener('updatefound', event => {
    waitForServiceWorkerActivation();
  });
}

async function waitForServiceWorkerActivation() {
  const registration = await navigator.serviceWorker.getRegistration(SCOPE);
  if (registration.active) {
    installPaymentHandlerInServiceWorker();
    return;
  }

  const serviceWorker = registration.installing || registration.waiting;
  serviceWorker.addEventListener('statechange', event => {
    if (event.target.state == 'activated') {
      installPaymentHandlerInServiceWorker();
    }
  });
}

async function installPaymentHandlerInServiceWorker() {
  const registration = await navigator.serviceWorker.getRegistration(SCOPE);
  await registration.paymentManager.instruments.clear();
  await registration.paymentManager.instruments.set('instrument-key', {
      name: 'Instrument Name',
      icons: [
        {src: '/images/rgrg-256x256.png', sizes: '256x256', type: 'image/png'},
      ],
      enabledMethods: ['basic-card'],
      capabilities: {supportedNetworks: ['mir'], supportedTypes: ['prepaid']},
  });
  runTests();
}

function runTests() {
  promise_test(async t => {
    const response = await new PaymentRequest(
      [{supportedMethods: 'basic-card'}],
      {
        id: 'test-payment-request-identifier',
        total: {label: 'Total', amount: {currency: 'USD', value: '0.01'}},
      },
    ).show();
    response.complete('success');
    assert_equals(response.requestId, 'test-payment-request-identifier');
    assert_equals(response.methodName, 'basic-card');
    assert_array_equals(response.details.billingAddress.addressLine, [
      '1875 Explorer St #1000',
    ]);
    assert_equals(response.details.billingAddress.city, 'Reston');
    assert_equals(response.details.billingAddress.country, 'US');
    assert_equals(response.details.billingAddress.dependentLocality, '');
    assert_equals(response.details.billingAddress.languageCode, '');
    assert_equals(response.details.billingAddress.organization, 'Google');
    assert_equals(response.details.billingAddress.phone, '+15555555555');
    assert_equals(response.details.billingAddress.postalCode, '20190');
    assert_equals(response.details.billingAddress.recipient, 'Jon Doe');
    assert_equals(response.details.billingAddress.region, 'VA');
    assert_equals(response.details.billingAddress.sortingCode, '');
    assert_equals(response.details.cardNumber, '4111111111111111');
    assert_equals(response.details.cardSecurityCode, '123');
    assert_equals(response.details.cardholderName, 'Jon Doe');
    assert_equals(response.details.expiryMonth, '12');
    assert_equals(response.details.expiryYear, '2028');
  }, 'Can perform payment');
}

installPaymentHandlerAndRunTheTests();
</script>
